Move Encoder class to its own module, as it does not need to be exposed to documentation
