+++ docs/source/encoders.rst
@@ -5,25 +5,8 @@
 :mod:`pwnlib.encoders` --- Encoding Shellcode
 ===============================================
 
+.. automodule:: pwnlib.encoders
+   :members:
+
 .. automodule:: pwnlib.encoders.encoder
-   :members:
-
-.. automodule:: pwnlib.encoders.i386.ascii_shellcode
-   :members:
-   :special-members:
-   :exclude-members: __init__
-
-.. automodule:: pwnlib.encoders.i386.xor
-   :members:
-
-.. automodule:: pwnlib.encoders.i386.delta
-   :members:
-
-.. automodule:: pwnlib.encoders.amd64.delta
-   :members:
-
-.. automodule:: pwnlib.encoders.arm.xor
-   :members:
-
-.. automodule:: pwnlib.encoders.mips.xor
    :members:
